requirements.txt
fastapi==0.92.0
numpy==1.24.2
pandas==1.5.3
pydantic==1.10.5
PyYAML==6.0
scikit_learn==1.2.1
uvicorn==0.20.0
